+++ src/utils/auth/createSignature.ts
