+++ src/utils/auth/extractAndVerifySignature.ts
