+++ src/utils/github/normalizeGithubUser.test.ts
