+++ src/utils/typing/arrays.test.ts
