+++ src/utils/typing/arrays.ts
