+++ test/utils/createSlackEvent.ts
@@ -1,7 +1,7 @@
 import { getSlackEvent } from '@test/utils/getSlackEvent';
 
 import { Fastify } from '@/types';
-import { createSignature } from '@utils/createSignature';
+import { createSignature } from '@/utils/auth/createSignature';
 
 function createSlackSignature(payload, timestamp) {
   return createSignature(
